refactor: 메서드 분리와 패키지 정리
