+++ mvc/src/main/java/nextstep/mvc/controller/tobe/HandlerExecution.java
@@ -7,15 +7,19 @@ import nextstep.mvc.view.ModelAndView;
 import java.lang.reflect.Method;
 
 public class HandlerExecution {
-    private Method method;
-    private Object execution;
+    private final Method method;
+    private final Object execution;
 
     public HandlerExecution(Method method, Object execution) {
         this.method = method;
         this.execution = execution;
     }
 
-    public ModelAndView handle(HttpServletRequest request, HttpServletResponse response) throws Exception {
-        return (ModelAndView) method.invoke(execution, request, response);
+    public ModelAndView handle(HttpServletRequest request, HttpServletResponse response) {
+        try {
+            return (ModelAndView) method.invoke(execution, request, response);
+        } catch (Exception e) {
+            throw new IllegalArgumentException(e.getMessage());
+        }
     }
 }
