+++ mvc/src/main/java/nextstep/mvc/servlet/AnnotationHandlerAdapter.java
@@ -1,12 +1,12 @@
-package nextstep.mvc;
+package nextstep.mvc.servlet;
 
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import nextstep.mvc.controller.asis.Controller;
-import nextstep.mvc.view.JspView;
+import nextstep.mvc.controller.tobe.HandlerExecution;
 import nextstep.mvc.view.ModelAndView;
 
-public class ManualHandlerAdapter implements HandlerAdapter{
+public class AnnotationHandlerAdapter implements HandlerAdapter {
     @Override
     public boolean supports(Object handler) {
         return handler instanceof Controller;
@@ -14,9 +14,7 @@ public class ManualHandlerAdapter implements HandlerAdapter{
 
     @Override
     public ModelAndView handle(HttpServletRequest request, HttpServletResponse response, Object handler) throws Exception {
-        final Controller controller = (Controller) handler;
-        final String viewName = controller.execute(request, response);
-        final JspView view = new JspView(viewName);
-        return new ModelAndView(view);
+        final HandlerExecution execution = (HandlerExecution) handler;
+        return execution.handle(request, response);
     }
 }
