+++ mvc/src/main/java/nextstep/mvc/servlet/ManualHandlerAdapter.java
@@ -1,12 +1,12 @@
-package nextstep.mvc;
+package nextstep.mvc.servlet;
 
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import nextstep.mvc.controller.asis.Controller;
-import nextstep.mvc.controller.tobe.HandlerExecution;
+import nextstep.mvc.view.JspView;
 import nextstep.mvc.view.ModelAndView;
 
-public class AnnotationHandlerAdapter implements HandlerAdapter {
+public class ManualHandlerAdapter implements HandlerAdapter {
     @Override
     public boolean supports(Object handler) {
         return handler instanceof Controller;
@@ -14,7 +14,9 @@ public class AnnotationHandlerAdapter implements HandlerAdapter {
 
     @Override
     public ModelAndView handle(HttpServletRequest request, HttpServletResponse response, Object handler) throws Exception {
-        final HandlerExecution execution = (HandlerExecution) handler;
-        return execution.handle(request, response);
+        final Controller controller = (Controller) handler;
+        final String viewName = controller.execute(request, response);
+
+        return new ModelAndView(new JspView(viewName));
     }
 }
